created a Symbolics extension
