+++ Project.toml
@@ -17,7 +17,6 @@ SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 SphericalFunctions = "af6d55de-b1f7-4743-b797-0829a72cf84e"
 SymbolicIndexingInterface = "2efcf032-c050-4f8e-a9bb-153293bab1f5"
 SymbolicUtils = "d1185830-fcd6-423d-90d6-eec64667417b"
-Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 FastDifferentiation = "eb9bf01b-bf85-4b60-bf87-ee5de06c00be"
 RuntimeGeneratedFunctions = "7e49a35a-f44a-4d26-94aa-eba1b4ca6b47"
 
@@ -39,3 +38,9 @@ Symbolics = "5.25"
 julia = "1.9"
 FastDifferentiation = "0.3.14"
 RuntimeGeneratedFunctions = "0.5.13"
+
+[extensions]
+PostNewtonianSymbolicsExt = "Symbolics"
+
+[weakdeps]
+Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
